==> 2. Get players.ipynb <==
from datetime import timedelta, date
import pandas
import requests
from bs4 import BeautifulSoup as bs
from pprint import pprint
#import re
import numpy as np
import string
from datetime import datetime, timedelta
from IPython.display import display, HTML

thisdir = '/home/idies/workspace/Storage/raddick/nba/'
teamfile = thisdir + 'teamlist.csv'

outdir = '/home/idies/workspace/Temporary/raddick/nba_scratch/'

def daterange(start_date, end_date):
    for n in range(int ((end_date - start_date).days)):
        yield start_date + timedelta(n)
print('ok')

all_players_df = pandas.DataFrame()
for ltr in string.ascii_lowercase:
    theurl = 'https://www.basketball-reference.com/players/{0:}/'.format(ltr)
    response = requests.get(theurl)
    if (response.status_code == 200):
        print('\t'+theurl)
        soup = bs(response.content, "html.parser", from_encoding="utf-8")
        thetables = soup.findAll('table')
        for thistable in thetables:
            tablebody = thistable.find('tbody')
            all_player_rows = tablebody.findAll('tr')
            player_info_records = []            
            for this_player_row in all_player_rows:
                this_player_info_dict = {}
                this_player_info_dict['playerID'] = this_player_row.find('th').get('data-append-csv')
                this_player_info_dict['player_name'] = this_player_row.find('th').get_text()
                for this_stat in this_player_row.findAll('td'):
                    this_player_info_dict[this_stat.get('data-stat')] = this_stat.get_text()
                player_info_records.append(this_player_info_dict)
        player_info_df = pandas.DataFrame.from_records(player_info_records)
        all_players_df = all_players_df.append(player_info_df)

print('fixing names...')
all_players_df = all_players_df.assign(name_first = all_players_df['player_name'].apply(lambda x: x[0:x.find(' ')]))
all_players_df = all_players_df.assign(name_last = all_players_df['player_name'].apply(lambda x: x[x.find(' ')+1:]))

all_players_df = all_players_df.assign(hall_of_fame = False)
all_players_df.loc[all_players_df['name_last'].apply(lambda x: x[-1] == '*'), 'hall_of_fame'] = True
all_players_df.loc[all_players_df['name_last'].apply(lambda x: x[-1] == '*'), 'name_last'] = all_players_df['name_last'][all_players_df['name_last'].apply(lambda x: x[-1] == '*')].apply(lambda x: x[:-1])



print('fixing positions...')
all_players_df = all_players_df.assign(pos_primary = all_players_df['pos'].apply(lambda x: x.split('-')[0]))
all_players_df = all_players_df.assign(pos_secondary = '')
all_players_df.loc[all_players_df['pos'].apply(lambda x: len(x.split('-')) > 1), 'pos_secondary'] = all_players_df['pos'][all_players_df['pos'].apply(lambda x: len(x.split('-')) > 1)].apply(lambda x: x.split('-')[1])

print('fixing career years and weight...')
for x in ['year_min', 'year_max', 'weight']:
    all_players_df.loc[:, x] = pandas.to_numeric(all_players_df[x], errors='coerce')

print('fixing height...')
all_players_df = all_players_df.assign(height_inches = all_players_df['height'].apply(lambda x: int(x.split('-')[0])*12 + int(x.split('-')[1])))

print('fixing birth date...')
all_players_df.loc[:, 'birth_date'] = pandas.to_datetime(all_players_df['birth_date'], errors='coerce')

print('fixing college(s)...')
##all_players_df['colleges'].apply(lambda x: len(x.split(','))).max()
all_players_df = all_players_df.assign(college = all_players_df['colleges'].apply(lambda x: x.split(',')[0]))
all_players_df = all_players_df.assign(college_second = '')
all_players_df.loc[all_players_df['colleges'].apply(lambda x: len(x.split(',')) > 1), 'college_second'] = all_players_df['colleges'][all_players_df['colleges'].apply(lambda x: len(x.split(',')) > 1)].apply(lambda x: x.split(',')[1])

print('fixing column order...')
all_players_df = all_players_df.set_index('playerID')
column_order = ['name_last', 'name_first', 'year_min', 'year_max', 'pos_primary', 'pos_secondary']
column_order += ['height_inches', 'weight', 'birth_date', 'college', 'college_second', 'hall_of_fame']
all_players_df = all_players_df[column_order]


all_players_df = all_players_df.sort_values(by='playerID')

all_players_df.to_csv(outdir+'players.csv', encoding='utf-8')

print('DONE!')
