==> 3. Combine data from all years.ipynb <==
import pandas
import os
thisdir = '/home/idies/workspace/Storage/raddick/nba/'

datadir = '/home/idies/workspace/Temporary/raddick/nba_scratch/'
print('ok')

#players_df = pandas.read_csv(datadir+'players.csv', encoding='utf-8', low_memory=False, index_col='playerID')
##players_df[players_df['name_last'] == 'Harden']#.groupby('playerID').size().sort_values(ascending=False).sort_index()
# James Harden playerID: hardeja01

#os.listdir(datadir)

#df1 = pandas.read_csv(datadir+'game_logs_2009_2010a.csv')
#df2 = pandas.read_csv(datadir+'game_logs_2009_2010b.csv')
#df3 = pandas.read_csv(datadir+'game_logs_2009_2010c.csv')
#df = pandas.concat((df1,df2,df3), axis=0)
#df = df.set_index(['playerID','gameID'])


#df.groupby('playerID')['pts'].sum()[df.groupby('playerID')['pts'].sum().index == 'hardeja01']

all_games_meta_df = pandas.DataFrame()
all_game_logs_df = pandas.DataFrame()

for i in range(2009, 2019):
    for j in ['a', 'b', 'c']:
        logs_df = pandas.read_csv(datadir+'game_logs_{0:.0f}_{1:.0f}{2:}.csv'.format(i, i+1, j), low_memory=False)
        #meta_df = pandas.read_csv(datadir+'games_meta_{0:.0f}_{1:.0f}{2:}.csv'.format(i, i+1, j), low_memory=False)
        #df = pandas.read_csv(datadir+'game_logs_{0:.0f}_{1:.0f}{2:}.csv'.format(i, i+1, j), low_memory=False)
        all_game_logs_df = pandas.concat((all_game_logs_df, logs_df), axis=0)
        #all_games_meta_df = pandas.concat((all_game_logs_df, meta_df), axis=0)
        
all_game_logs_df

.
